Expanded Support and Clarity
ACE:
Air/Ground, Invin, and Goto commands are now labeled in the anmchr command in realtime, (1_2F, 1_3C, 0_01, etc. more to come)
SE:
Updated labeling for cmdcombo and cmdspatk to be correct for todays standards (removed a lot of "?" and guessing).
DHC entries in CSP are now labeled as DHC instead of Either
PP is now replaced with AA, Atk+S is now labeled as AS